image_generation/requirements.txt
tensorflow
tensorflow-gpu
tensorflow-hub
numpy
Pillow
scipy
Flask
gunicorn